+++ NEWS.md
@@ -1,3 +1,5 @@
 # metafiler 0.1.2
 
 * Added a `NEWS.md` file to track changes to the package.
+* Added `cluster_samples()` and `cluster_features()` for reordering based on 
+agglomerative hierchical clustering